requirements.txt
fastapi==0.112.1
fastapi-cli==0.0.5
SQLAlchemy==2.0.32
python-jose==3.3.0
passlib==1.7.4
pydantic-settings==2.4.0
PyMySQL==1.1.1
cryptography==43.0.0